Respuesta del Delete

+++ demo/src/main/java/com/project/demo/rest/categoria/CategoriaRestController.java
@@ -1,9 +1,12 @@
 package com.project.demo.rest.categoria;
 
 import com.project.demo.logic.entity.categoria.Categoria;
+import com.project.demo.logic.entity.producto.Producto;
 import com.project.demo.logic.entity.user.User;
 import com.project.demo.logic.entity.categoria.CategoriaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -11,6 +14,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/categorias")
@@ -57,10 +61,22 @@ public class CategoriaRestController {
                 });
     }
 
+//    @DeleteMapping("/{id}")
+//    @PreAuthorize("hasAnyRole( 'SUPER_ADMIN_ROLE')")
+//    public void deleteCategoria(@PathVariable Long id) {
+//        CategoriaRepository.deleteById(id);
+//    }
     @DeleteMapping("/{id}")
-    @PreAuthorize("hasAnyRole( 'SUPER_ADMIN_ROLE')")
-    public void deleteCategoria(@PathVariable Long id) {
-        CategoriaRepository.deleteById(id);
+    @PreAuthorize("hasAnyRole('SUPER_ADMIN_ROLE')")
+    public ResponseEntity<String> deleteCategoria(@PathVariable Long id) {
+        Optional<Categoria> productoOptional = CategoriaRepository.findById(id);
+
+        if (((Optional<?>) productoOptional).isPresent()) {
+            CategoriaRepository.deleteById(id);
+            return new ResponseEntity<>("Categoria borrada con éxito", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("Categoria no encontrado", HttpStatus.NOT_FOUND);
+        }
     }
 
     @GetMapping("/me")
