+++ demo/src/main/java/com/project/demo/rest/productos/ProductosRestController.java
@@ -3,6 +3,8 @@ package com.project.demo.rest.productos;
 import com.project.demo.logic.entity.producto.Producto;
 import com.project.demo.logic.entity.producto.ProductoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -10,6 +12,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/productos")
@@ -61,8 +64,15 @@ public class ProductosRestController {
 
     @DeleteMapping("/{id}")
     @PreAuthorize("hasAnyRole('SUPER_ADMIN_ROLE')")
-    public void deleteProducto(@PathVariable Long id) {
-        ProductoRepository.deleteById(id);
+    public ResponseEntity<String> deleteProducto(@PathVariable Long id) {
+        Optional<Producto> productoOptional = ProductoRepository.findById(id);
+
+        if (((Optional<?>) productoOptional).isPresent()) {
+            ProductoRepository.deleteById(id);
+            return new ResponseEntity<>("Producto borrado con éxito", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("Producto no encontrado", HttpStatus.NOT_FOUND);
+        }
     }
 
     @GetMapping("/me")
